Refactor to view helper
