+++ app/controllers/huawei_router_passwords_controller.rb
@@ -3,18 +3,5 @@ class HuaweiRouterPasswordsController < ApplicationController
 
   def new
     @password = Settings.huawei.devices.password
-    @breadcrumb_path = if current_user.is_responsible_body_user?
-                         responsible_body_internet_path
-                       elsif current_user.is_school_user?
-                         internet_school_path(current_user.school)
-                       else
-                         safe_path_for_other_user_types
-                       end
-  end
-
-private
-
-  def safe_path_for_other_user_types
-    root_path
   end
 end
