+++ app/views/huawei_router_passwords/new.html.erb
@@ -2,7 +2,7 @@
 <% content_for :title, title %>
 <%- content_for :before_content do %>
   <% breadcrumbs([{ "Home" => root_path },
-                  { t('page_titles.signed_in_internet_access') => @breadcrumb_path },
+                  { t('page_titles.signed_in_internet_access') => last_breadcrumb_path_for_huawei },
                   title,
                  ]) %>
 <% end %>
